chore: extracted url api into separate file
